pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wagtail-wordpress-image-importer"
version = "0.1.0"
description = "Import WordPress images into Wagtail"
readme = "README.md"
authors = [{name = "Tom Dyson"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Wagtail",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["wagtail", "wordpress", "images", "import"]
dependencies = [
    "wagtail>=5.0",
    "requests>=2.25.0",
    "Pillow>=9.0.0",
    "tqdm>=4.65.0",
]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/tomdyson/wagtail-wordpress-image-importer"
